Change Shooter Motor Name
Changed shootLeftMotor to be particleMotor. Changed shootRightMotor to be shooterMotor. Changed comments on if statements to be more easily understood. Added a variable for particle motor for the telemetry. Added particle readout to telemetry. Added (Servo Code) in comment next to beginning of code that is used for servos.

This will help to make the code easier to read. I will change the other file to account for changes.