Change Names of Shooter Motors to Work with teleop_CS edits
Changed "shootLeftMotor" to "particeMotor" and "shootRightMotor" to "shooterMotor".  Changed the names again in the initialization section. Changed "particleMotor" configuration name to be "Particle Motor" and "shooterMotor" to be "Shooter Motor". Changed names again in "Set all motors to zero power" section. Added (Servo Code) to comments where servo code is.

These edits will help with understanding the code and fixing the code to work with the teleop_CS edits made.